move & rename DataQueueHandler subscription manager to better place for using in ToolBox
