+++ Common/Data/EventBasedDataQueueHandlerSubscriptionManager.cs
@@ -1,0 +1,44 @@
+﻿/*
+ * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
+ * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+*/
+
+
+using System;
+using System.Collections.Generic;
+
+namespace QuantConnect.Data
+{
+    public class EventBasedDataQueueHandlerSubscriptionManager : DataQueueHandlerSubscriptionManager
+    {
+        public Func<IEnumerable<Symbol>, TickType, bool> SubscribeImpl;
+        public Func<IEnumerable<Symbol>, TickType, bool> UnsubscribeImpl;
+
+        public Func<TickType, string> GetChannelName;
+
+        protected override bool Subscribe(IEnumerable<Symbol> symbols, TickType tickType)
+        {
+            return SubscribeImpl?.Invoke(symbols, tickType) == true;
+        }
+
+        protected override bool Unsubscribe(IEnumerable<Symbol> symbols, TickType tickType)
+        {
+            return UnsubscribeImpl?.Invoke(symbols, tickType) == true;
+        }
+
+        protected override string ChannelNameFromTickType(TickType tickType)
+        {
+            return GetChannelName?.Invoke(tickType);
+        }
+    }
+}
